requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==4.5.3
cryptography==2.7
Django==2.2.2
django-extensions==2.1.9
django-filter==2.1.0
django-money==0.15
django-unixdatetimefield==1.0.3
djangorestframework==3.9.4
gunicorn==19.9.0
httpretty==0.9.6
idna==2.8
jsonschema==3.0.1
psycopg2==2.8.3
py-moneyed==0.8.0
pycparser==2.19
pyOpenSSL==19.0.0
pyrsistent==0.15.2
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
Werkzeug==0.15.4
